Fix file descriptor leak when archiving test results
The process of copying the remote test results files to Jenkins for
archiving leaves a file stream open. This causes a build-up of open file
descriptors on the Jenkins Master which will eventually crash a unix
system depending on ulimit 'open files' setting.

This fix replaces the inline copying of files with the build-in
FilePath.copyFrom() method which closes the streams correctly.